pyproject.toml
[project]
name = "tg-bot-recommender-backend"
version = "0.1.0"
description = ""
authors = [
    { name = "domster704", email = "domster704@mail.ru" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.116.2,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
]

[tool.poetry]
packages = [{ include = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "black (>=25.11.0,<26.0.0)"
]
